prepare_host: Fix configuration of public network and host ip
The default standalone configuration seems to want to configure a
single interface. Rather than fight it, this change moves us to a
single dummy interface.

We only configure the public IP on this interface. The control plane
IP is configured by tripleo standalone during the deploy.

We also add a route to the public network. It seems this has to be
added to the bridge rather than the dummy interface, which means we
can't add it until the bridge is up. We add a systemd unit with
infinite retry to help us here.