Enable Octavia
Enable Octavia (Load Balancer as a Service in OpenStack) by default,
since it's used when enabling Kuryr instead of OpenShiftSDN.

Note that we have to specify both OctaviaGenerateCerts and
OctaviaCaKeyPassphrase, to make the deployment happy but since we don't
use the Amphora driver (we use OVN-provider), it doesn't matter that the
passphrase is not secured.
We also need to create a keypair to make the installer happy but since
we don't use the Amphora, this isn't needed in our ShiftStack use case.

Signed-off-by: Emilien Macchi <emilien@redhat.com>