Use tripleo operator to deploy standalone
Instead of running a shell script from systemd, let's rely on the
TripleO Operator that runs the standalone deployment with a lot of
options:
https://docs.openstack.org/tripleo-operator-ansible/latest/roles/role-tripleo_deploy.html

We'll be able to override these options for free and just consume that
role which is heavily tested by TripleO CI and used by many.

Note: the octavia SSH key is now in its own Ansible task to assume
compatibility.