Add support to install Ceph via loop device
For the developer use case, we use a loop device to install LVM volumes
for ceph.
This patch will add two options to dev-install:

* ceph_enabled (false by default)
* ceph_loop_device_size (10 GB by default)

It will take care of creating a loop device that is reboot-persistent,
and then use it for the deployment of Ceph via ceph-ansible.
It'll also dynamically load the environment and services if we want
Ceph enabled.

This addition does not configure Ceph for production, so in the
OpenShift on OpenStack use-case, keep in mind that the Ceph
performances won't be great on loop devices. In other words, booting
the cluster nodes from volumes will be problematic because of perf
requirements for etcd.

It also adds nodisk flavors for both large and xlarge. They can be
useful when deploying OCP masters or workers with rootVolume from
Cinder (backend in Ceph).