Allow to create tunnels between virtual bridges
For both hostonly (used as a provider network) and control plane
bridges, we now allow them to be plugged to remote hosts bridges via
VXLAN tunnels.

The new parameter is `tunnel_remote_ips` (default: []), where for each
IP in the list, we'll create a unique tun interface, connect it to the
bridge on a specific VNI (100 for control plane and 101 for hostonly).

Use-cases:
* for control plane, in the case of DCN the AZ nodes need to reach the
  central node control plane.
* for provider network, the workloads plugged on this network might
  want to reach another workload running on another hypervisor,
  therefore it makes sense to plug the provider networks together.