Switch bridge MTU to 1400 if tunnels are being used.
When using VXLAN tunnels between the bridges, it's important to set proper MTU because the
Ethernet frame is being encapsulated in `Ethernet + IP + UDP + Overlay Header`.
This means that the Maximum Transmission Unit for the underlay needs to be adjusted to something close to 1400.
Using 1400 is a safe number for now which should give up flexibility in
using IPv4 and IPv6 in the future.