manila: create the share type with snapshot support when possible
The snapshot support was added in recent version of OpenStack (current
master and backported to Wallaby).
For OSP16, we want to set it to False like it was the case before that
patch. But for new versions, we need to set it to True otherwise the
Manila Scheduler will fail to create shares:
http://paste.openstack.org/show/ivBPVOf9IcEOpUI9CfSL/