Add support for cephadm instead of ceph-ansible
Ceph has given up ceph-ansible and therefore TripleO had to remove its
support going forward.

This patch will detect if ceph-ansible can still be used and if not
it'll configure ceph using `cephadm`.

From a user standpoint, dev-install still offers 3 options:

* Disable the management of Ceph from dev-install; if users don't want
  Ceph or if they want to use `extra_heat_params`.
* Let cephadm to install Ceph OSDs on available disks that aren't
  rotational (for performance purposes). This is different from
  ceph-ansible where we could choose the disks, but one can achieve it
  via option #1 if they don't have SSD/NVME disks.
* Let dev-install creating LVM resources on a loop device and let
  cephadm to use these volumes to configure Ceph. Note that cephadm
  doesn't need the DB volume but this shouldn't matter anyway, as the
  data volume should be big enough to run the workloads (volumes, vms,
  etc), since it's 96% of the total size of the loop device.

In other words, this change is backward compatible except if you
deployed Ceph with `ceph_devices` that are rotational; then no OSD will
be deployed and cephadm will crash. In that case it's suggested to go
with option #1 previously described, but dev-install maintainers don't
suggest to install Ceph on these disks because of performance impact.