Configure MTU on dummy interface, not the bridge
The bridge takes the highest MTU of the interfaces connected to it.
So it was always set to 1500 even if we force it to 1400, because dummy
nics were using 1500.

Now that we set it to 1400, the bridge will be 1400 too, and it helps
when setting up VXLAN tunnels in multinode environment.

This has been tested against a real environment.