Configure Cinder API to allow users to set volume states
By default, volume states can only be changed by admin but in
dev-install cases, we are operating our cloud and sometimes need to
reset the state of a volume.

e.g. in OpenShift CI, sometimes volumes aren't cleaned correctly and we
have to manually clear them up. Thanks to this patch, the tenant will
be able to set the state of their own volumes.