Improve NFV support
Adding a new parameter, `dpdk_interface` which allows to enable DPDK in
dev-install.

It'll do the following:
* Pre-configure the Kernel before the configuration of DPDK (and reboot
  if necessary)
* Configure Tuned before installing OpenStack
* Configure TripleO roles & services, to enable basic parameters for
  OVN or OVS DPDK.
* Create a specific network_config when DPDK is enabled, and only use
  OVS for the DPDK network; and leave the other networks on their own
  interface instead of a bridge. The dummy interfaces will be configured
  via nmstate, with their proper IP config.
* Dynamically configure Neutron Bridge mapping for DPDK.
* Create flavors for DPDK, they're optional but can be useful.
* Manage the provider network dedicated to DPDK and deal with SNAT like
  we did for hostonly already.
* Set NovaComputeStartupDelay: to 180, so after a reboot it'll wait
  2 min before starting nova-compute. It will give some time to Ceph to
  start and avoid errors in Nova.

Some changes for SR-IOV:
* Create a dedicated SR-IOV network when needed

When SR-IOV is enabled, we will create a dedicated provider network
where the SR-IOV nic will be plugged into.

This will allow to separate the provider networks from regular
interfaces, SR-IOV and later DPDK.

It won't do the following:
* Configure CPU partitioning for Nova, the host and DPDK
However it provides an example of overrides that can be used to deploy
with DPDK.

Note: we had to partially import Ansible roles from tripleo-ansible.
The reason is that we have to deal with Kernel & tuned before
installing OpenStack otherwise the deployment of OpenStack will fail at
the OVS DPDK step, since the kernel wouldn't have been configured and
the host rebooted. So we imported the roles, run it in `prepare_host`
step so when TripleO is being run, the Kernel & Tuned are already
configured.