SRIOV: enable DHCP agent
The DHCP agent is still needed if we need the VMs to get an IP when
booting from "direct" type port. In OSP17, OVN will be able to provide
DHCP directly but this wasn't bacported yet to OSP16.

This patch will deploy the Neutron DHCP agent if SRIOV is being
configured on at least one interface. Note that the agent won't be used
by anything else than the SRIOV nic agent, since OVN provide DHCP for
regular ports.