Remove management of RHCOS images
This has existed since the introduction of the 'prepare_stack' target
and corresponding playbook. However it's not strictly necessary for
non-CI use cases and may in fact be undesirable even in these cases, as
commit 8bd6a64a suggests. To top things off, it's actually broken now
since the installer removed the 'data/data/rhcos.json' file as part of
OpenShift 4.10 [1]. Remove the entire feature and related config knobs.

[1] https://github.com/openshift/installer/commit/d773ee557

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>