Simplify DPDK integration with provider network
Re-use `hostonly` provider network when DPDK is enabled and instead of
adding a dummy interface (dummy1), just use the physical interface that
we want to use for DPDK.

This will avoid confusions when our users see multiple provider
networks.
In fact, one is enough and should be used.

Note: we maintain the SRIOV provider network separated, since it's using
another kind of interface etc.