Handle boolean 'ceph_devices'
In OSP17, Ceph deployment is done via the 'ceph-adm' tool (native to
ceph tooling) which doesn't allow us to specify a list of disk devices
to use. Instead, the tool selects which disk to use for Ceph based on
its properties. As a result, the 'ceph_devices' variable must be a
boolean for OSP 17 deployments, rather than the list used in OSP 16.
Handle this.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>