install_stack: disable the use of --standalone arg in recent TripleO
Tripleoclient removed --standalone option some time ago:
https://review.opendev.org/c/openstack/python-tripleoclient/+/753447

And the TripleO Operator Ansible was updated to not use it anymore:
https://review.opendev.org/c/openstack/tripleo-operator-ansible/+/753853

In dev-install, we now figure out if the version of TripleO that is
being installed on CentOS needs or doesn't need that flag.

Note that we'll have to figure out how to make it work for OSP 17.1 and
OSP 18, since they'll probably contain these changes, but we have time
for now.