IPv6 Dual Stack on hostonly network
From now, the `hostonly` network will be configured in dual stack so we
can easily play with IPv6 on that provider network.

We'll use 2001:db8::/64 cidr and make sure that we can create VMs on
that network, and ping from each other and even reach outside out of the
box.

Tested on OSP 16.2.