Align Ceph deployment to OSP 17
This patch adds a new Ceph role which is responsible to execute a set of
tasks and deploy the cluster in the standalone scenario. This works
aligns dev-install to the new approach introduced in Wallaby/Zed, where
the Ceph bootstrap is not part of the overcloud deploy anymore.

In particular, as per [1], Ceph expects to see:
1. an already provisioned metal
2. an already provisioned storage_network
3. reserved VIPs (if needed)

There are still actions happening during the overcloud deployment (pool
creation, ganesha deployment) to finalize the cloud, but the initial
bootstrap process has been moved to a set of overcloud ceph related
commands, executed through the related tripleo.operator.

Note that a similar approach has been recently introduced in oooq [2].

[1] https://docs.openstack.org/project-deploy-guide/tripleo-docs/latest/features/deployed_ceph.html#deployed-ceph-workflow
[2] https://github.com/openstack/tripleo-quickstart-extras/blob/master/roles/standalone/tasks/ceph-install.yml

Signed-off-by: Francesco Pantano <fpantano@redhat.com>
Co-Authored-By: Emilien Macchi <emilien@redhat.com>