CinderApiPolicies: allow force-detach
We have observed in our CI that sometimes volumes aren't detached when
being removed.
This can be due to the density of testing in our environment.
To address this, we allow tenants to force-detach their own volumes.

In our CI scripts, we'll use that to force detach volumes before they
are cleaned-up in the periodic cleanup jobs.